requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.11

# Configuration
django-environ==0.4.0



# Forms
django-braces==1.9.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.5.2

# Images
Pillow==3.3.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.31.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.4
redis>=2.10.0


celery==3.1.23


rcssmin==1.0.6 
django-compressor==2.1


# Your custom requirements go here

django-mptt==0.8.6
django-jsonify==0.3.0
django-simple-history==1.8.1
# django-easy-select2==1.3.3
djangorestframework==3.6.2
django-rest-swagger==2.1.2
drf-extensions==0.3.1
markdown==2.6.8
django-rest-auth==0.9.0
django-filter==1.0.2
beautifulsoup4==4.5.3
#google-api-python-client==1.6.2
aiohttp==2.0.3
channels==1.1.3
asgi-redis==1.3.0